move print_rel_notes to a releasing tree and make part of distro
